Query works with the store.
Note due to an error in the translator the cdao namespace is
http://www.evolutonaryontology.org/cdao.owl# 
rather than
http://www.evolutionaryontology.org/cdao.owl#
for most of the entries. This issue has been addressed and the changes
are being merged. For now though use the typo versoin to get the complete set.
(It takes a week or two to merge all the trees back into the store).

Note also I haven't yet addressed the initialization of the cdao namespace in
the dict so to use it for example for a node type the literal as follows
<http://www.evolutionaryontology.org/cdao.owl#Node> for nodes. I'm working on this
so hopefully soon the cdao:Node version will work *fingers crossed*