Developed a heap-like memory manager for allocating parts of shared segments. Started work on a flex/bison version of the tree parser to take care of the 6-8 cases that don't work currently, and be easier to maintain than the hand-written lexer/parser of the current implementation.
Added some descriptions to the triplestore page, and did some reorganizing suggested by Arlin.
Started adapting node to allocate it's objects into a shared segment space rather than the heap.
This will allow the flex/bison implementation to read the string data from a pipe, and put the 
resulting parse tree in a shared space for the rest of the translator to use.